Fix: User ID 타입 변경에 따른 Chat 관련 로직 수정

+++ Back/petPlace/src/main/java/com/minjeok4go/petplace/chat/service/ChatRoomService.java
@@ -29,7 +29,7 @@ public class ChatRoomService {
         this.userChatRoomService = userChatRoomService;
     }
 
-    public ChatRoomDTO createChatRoom(Integer userId1, Integer userId2) {
+    public ChatRoomDTO createChatRoom(Long userId1, Long userId2) {
         User user1 = userRepository.findById(userId1).orElseThrow();
         User user2 = userRepository.findById(userId2).orElseThrow();
 
@@ -60,7 +60,7 @@ public class ChatRoomService {
     }
 
 
-    public List<ChatRoomDTO> getChatRoomsByUser(Integer userId) {
+    public List<ChatRoomDTO> getChatRoomsByUser(Long userId) {
         List<ChatRoom> rooms = chatRoomRepository.findByUserId(userId);
         // ChatRoom → ChatRoomDto 변환
         return rooms.stream()
